split responses into smaller files
