+++ server/src/controllers/users/signUp.controller.ts
@@ -7,7 +7,8 @@ import { TUserSchema, userSchema } from "../../dtos/user.dto";
 import bcrypt from "bcryptjs"
 import { fromZodError } from "zod-validation-error"
 import { messages } from "../../messages";
-import { TMainResponse, TSignUpResponse, } from "../../types/responses";
+import { TMainResponse } from "../../types/responses";
+import { TSignUpResponse } from "../../types/responses/user.response";
 
 export async function signUpController(req: Request<{}, {}, TUserSchema>, res: Response<TMainResponse<TSignUpResponse>>, next: NextFunction) {
     const { username, name, password } = req.body
