+++ server/src/types/responses/post.response.ts
@@ -1,0 +1,17 @@
+import { TPost } from "../models"
+
+export type TCreatePostResponse = {
+    post: TPost
+}
+
+export type TUpdatePostResponse = {
+    post: TPost
+}
+
+export type TDeletePostResponse = {
+    post: TPost
+}
+
+export type TGetSinglePostResponse = {
+    post: TPost
+}
